requirements.txt
Flask==2.3.3
gspread==5.12.0
oauth2client==4.1.3
python-dotenv==1.0.1
